// src/lib/database/types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type Database = {
    public: {
        Tables: {
// ...
            usage: {
                Row: {
                    current_quota: number | null;
                    data: Json | null;
                    is_subscription_canceled: boolean | null;
                    is_subscription_paused: boolean | null;
                    last_reset_date: string;
                    last_transaction_id: string | null;
                    management_url: string | null;
                    next_reset_date: string;
                    plan_type: Database['public']['Enums']['subscription_plan'];
                    premium_expires_date: string | null;
                    premium_grace_period_expires_date: string | null;
                    premium_product_identifier: string | null;
                    premium_product_plan_identifier: string | null;
                    premium_purchase_date: string | null;
                    status: Database['public']['Enums']['subscription_status'];
                    store: Database['public']['Enums']['store_type'] | null;
                    user_id: string;
                };
                Insert: {
                    current_quota?: number | null;
                    data?: Json | null;
                    is_subscription_canceled?: boolean | null;
                    is_subscription_paused?: boolean | null;
                    last_reset_date: string;
                    last_transaction_id?: string | null;
                    management_url?: string | null;
                    next_reset_date: string;
                    plan_type: Database['public']['Enums']['subscription_plan'];
                    premium_expires_date?: string | null;
                    premium_grace_period_expires_date?: string | null;
                    premium_product_identifier?: string | null;
                    premium_product_plan_identifier?: string | null;
                    premium_purchase_date?: string | null;
                    status: Database['public']['Enums']['subscription_status'];
                    store?: Database['public']['Enums']['store_type'] | null;
                    user_id?: string;
                };
                Update: {
                    current_quota?: number | null;
                    data?: Json | null;
                    is_subscription_canceled?: boolean | null;
                    is_subscription_paused?: boolean | null;
                    last_reset_date?: string;
                    last_transaction_id?: string | null;
                    management_url?: string | null;
                    next_reset_date?: string;
                    plan_type?: Database['public']['Enums']['subscription_plan'];
                    premium_expires_date?: string | null;
                    premium_grace_period_expires_date?: string | null;
                    premium_product_identifier?: string | null;
                    premium_product_plan_identifier?: string | null;
                    premium_purchase_date?: string | null;
                    status?: Database['public']['Enums']['subscription_status'];
                    store?: Database['public']['Enums']['store_type'] | null;
                    user_id?: string;
                };
                Relationships: [];
            };
// ...
        };
// ...
    };
};

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>;

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, 'public'>];
// ...
export type Enums<
    DefaultSchemaEnumNameOrOptions extends
        | keyof DefaultSchema['Enums']
        | { schema: keyof DatabaseWithoutInternals },
    EnumName extends DefaultSchemaEnumNameOrOptions extends {
        schema: keyof DatabaseWithoutInternals;
    }
        ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
        : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
}
    ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
    : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
      ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
      : never;